Move to supporting 10.10; Use DiffFormatterApp and framework tests
