+++ DiffFormatterApp/DiffFormatterApp.h
@@ -1,0 +1,19 @@
+//
+//  DiffFormatterApp.h
+//  DiffFormatterApp
+//
+//  Created by Dan Loman on 1/13/19.
+//  Copyright © 2019 DHL. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+
+//! Project version number for DiffFormatterApp.
+FOUNDATION_EXPORT double DiffFormatterAppVersionNumber;
+
+//! Project version string for DiffFormatterApp.
+FOUNDATION_EXPORT const unsigned char DiffFormatterAppVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <DiffFormatterApp/PublicHeader.h>
+
+
