+++ DiffFormatterApp/Extensions/Sequence+DiffFormatter.swift
