+++ DiffFormatterApp/Output/OutputGenerator.swift
