+++ DiffFormatterAppTests/Config/excluded_section_config.json
